Fixed the percentages

+++ main.cpp
@@ -32,9 +32,9 @@ int main(void) {
 void results() {
 
     clearScreen();
-	int total = positive + negative;
-	int negativePros = (double)negative / (double)total * 100.0;
-	int positivePros = (double)positive / (double)total * 100.0;
+	double total = positive + negative;
+	double negativePros = (double)negative / (double)total * 100.0;
+	double positivePros = (double)positive / (double)total * 100.0;
 	cout << "You entered the following integers:" << endl;
 	cout << "Positive numbers: " << positive << endl;
 	cout << "Positive number percentage: " << positivePros << "%" << endl;
